Remove now-defunct translate/ folder and fix spellcheck loading
