+++ app/components/Dialogs.tsx
@@ -15,8 +15,6 @@ import {QuestType, ShareType, DialogsState, DialogIDType} from '../reducers/Stat
 import theme from '../theme'
 
 declare var ga: any;
-var MarkdownParserError: any = (require('../../translation/to_xml') as any).MarkdownParserError;
-var XMLParserError: any = (require('../../translation/to_markdown') as any).XMLParserError;
 
 interface ErrorDialogProps extends React.Props<any> {
   open: boolean;
@@ -31,15 +29,6 @@ export class ErrorDialog extends React.Component<ErrorDialogProps, {}> {
       var error = this.props.errors[i];
       ga('send', 'event', 'error', error.name, error.message);
       console.log(error.stack);
-
-      if (error instanceof MarkdownParserError || error instanceof XMLParserError) {
-        errors.push(<div key={i}>
-          <strong>{error.name}: "{error.message}"</strong>
-          <div><strong>Line:</strong> {error.line}</div>
-          <div><strong>Usage:</strong> {error.usage}</div>
-        </div>);
-        continue;
-      }
       errors.push(<div key={i}>{error.toString()}</div>);
     }
