+++ app/components/MainContainer.tsx
@@ -4,8 +4,6 @@ import {setDirty} from '../actions/editor'
 import {AppState} from '../reducers/StateTypes'
 import Main, {MainStateProps, MainDispatchProps} from './Main'
 
-var toMarkdown: any = require('../../translation/to_markdown')
-
 const mapStateToProps = (state: AppState, ownProps: any): MainStateProps => {
   return {
     loggedIn: state.user.loggedIn,
