+++ app/components/QuestIDEContainer.tsx
@@ -6,8 +6,6 @@ import {AppState, QuestType, EditorState} from '../reducers/StateTypes'
 import QuestIDE, {QuestIDEStateProps, QuestIDEDispatchProps} from './QuestIDE'
 import {store} from '../store'
 
-var toMarkdown: any = require('../../translation/to_markdown')
-
 const mapStateToProps = (state: AppState, ownProps: any): QuestIDEStateProps => {
   return {
     editor: state.editor,
