--- translation/to_markdown.d.ts
@@ -1,4 +1,0 @@
-declare module "to_markdown" {
-  export var toMarkdown: any;
-  export var XMLParserError: any;
-}
