--- translation/to_xml.d.ts
@@ -1,4 +1,0 @@
-declare module "to_xml" {
-  export var toXML: any;
-  export var MarkdownParserError: any;
-}
